+++ src/main/java/com/github/jlangch/venice/impl/types/collections/VncSequence.java
@@ -1,6 +1,7 @@
 package com.github.jlangch.venice.impl.types.collections;
 
 import java.util.List;
+import java.util.stream.Stream;
 
 import com.github.jlangch.venice.impl.types.VncVal;
 
@@ -9,4 +10,7 @@ public abstract class VncSequence extends VncCollection {
 
 	public abstract List<VncVal> getList();
 
+	public  Stream<VncVal> stream() {
+		return getList().stream();
+	}
 }