+++ ChangeLog.md
@@ -11,7 +11,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ### Added
 
-- added function assoc-in and get-in
+- added function merge, assoc-in, and get-in
 - added function update
 - added function vec
 - added function every? and any? operating on sequential collections